+++ examples/IN_EfficientFormer.py
@@ -4,13 +4,15 @@
 # LICENSE file in the root directory of this source tree.
 
 
+from pathlib import Path
+
 import pytorch_lightning as pl
 import torch
-from pl_bolts.datamodules import CIFAR10DataModule
+from pl_bolts.datamodules import ImagenetDataModule
 from torch import nn
 from torchmetrics import Accuracy
 
-from examples.cifarViT import Classifier, VisionTransformer
+from examples.cifar_ViT import Classifier, VisionTransformer
 from xformers.components import MultiHeadDispatch
 from xformers.components.attention import ScaledDotProduct
 from xformers.components.patch_embedding import PatchEmbeddingConfig  # noqa
@@ -49,16 +51,7 @@ class EfficientFormer(VisionTransformer):
         # but a pooling layer has been removed, due to the very small image dimensions in Cifar (32x32)
         base_hierarchical_configs = [
             BasicLayerConfig(
-                embedding=64,
-                attention_mechanism="pooling",
-                patch_size=3,
-                stride=2,
-                padding=1,
-                seq_len=image_size * image_size // 4,
-                feedforward="Conv2DFeedforward",
-            ),
-            BasicLayerConfig(
-                embedding=128,
+                embedding=48,
                 attention_mechanism="pooling",
                 patch_size=3,
                 stride=2,
@@ -67,7 +60,7 @@ class EfficientFormer(VisionTransformer):
                 feedforward="Conv2DFeedforward",
             ),
             BasicLayerConfig(
-                embedding=320,
+                embedding=96,
                 attention_mechanism="pooling",
                 patch_size=3,
                 stride=2,
@@ -75,8 +68,24 @@ class EfficientFormer(VisionTransformer):
                 seq_len=image_size * image_size // 64,
                 feedforward="Conv2DFeedforward",
             ),
-            # L1 would have an extra layer here, similar to the above,
-            # bringing the sequence length down to HxW / 1024
+            BasicLayerConfig(
+                embedding=224,
+                attention_mechanism="pooling",
+                patch_size=3,
+                stride=2,
+                padding=1,
+                seq_len=image_size * image_size // 256,
+                feedforward="Conv2DFeedforward",
+            ),
+            BasicLayerConfig(
+                embedding=448,
+                attention_mechanism="pooling",
+                patch_size=3,
+                stride=2,
+                padding=1,
+                seq_len=image_size * image_size // 1024,
+                feedforward="Conv2DFeedforward",
+            ),
         ]
 
         # Fill in the gaps in the config
@@ -85,7 +94,7 @@ class EfficientFormer(VisionTransformer):
             layernorm_style="pre",
             use_rotary_embeddings=False,
             mlp_multiplier=4,
-            in_channels=3,  # 24 if L1, there's another stem prior to the trunk
+            in_channels=24,  # 24 if L1, there's a stem prior to the trunk
         )
 
         # Now instantiate the EfficientFormer trunk
@@ -94,17 +103,16 @@ class EfficientFormer(VisionTransformer):
 
         self.trunk = xFormer.from_config(config)
 
-        # L1 model
-        # # This model requires a pre-stem (a conv prior to going through all the layers above)
-        # self.pre_stem = build_patch_embedding(
-        #     PatchEmbeddingConfig(
-        #         in_channels=3, out_channels=24, kernel_size=3, stride=2, padding=1
-        #     )
-        # )
+        # This model requires a pre-stem (a conv prior to going through all the layers above)
+        self.pre_stem = build_patch_embedding(
+            PatchEmbeddingConfig(
+                in_channels=3, out_channels=24, kernel_size=3, stride=2, padding=1
+            )
+        )
 
         # This model requires a final Attention step
         self.attention = MultiHeadDispatch(
-            dim_model=320, num_heads=4, attention=ScaledDotProduct()
+            dim_model=448, num_heads=8, attention=ScaledDotProduct()
         )
 
         # The classifier head
@@ -116,7 +124,7 @@ class EfficientFormer(VisionTransformer):
 
     def forward(self, x):
         x = x.flatten(-2, -1).transpose(-1, -2)  # BCHW to BSE
-        # x = self.pre_stem(x) # L1 model
+        x = self.pre_stem(x)
         x = self.trunk(x)
         x = self.attention(x)
         x = self.ln(x)
@@ -131,8 +139,8 @@ if __name__ == "__main__":
 
     # Adjust batch depending on the available memory on your machine.
     # You can also use reversible layers to save memory
-    REF_BATCH = 768
-    BATCH = 768  # lower if not enough GPU memory
+    REF_BATCH = 128
+    BATCH = 32  # lower if not enough GPU memory
 
     MAX_EPOCHS = 50
     NUM_WORKERS = 4
@@ -144,8 +152,11 @@ if __name__ == "__main__":
     # We'll use a datamodule here, which already handles dataset/dataloader/sampler
     # See https://pytorchlightning.github.io/lightning-tutorials/notebooks/lightning_examples/cifar10-baseline.html
     # for a full tutorial
-    dm = CIFAR10DataModule(
-        data_dir="data",
+    dm = ImagenetDataModule(
+        data_dir=Path.home()
+        / Path(
+            "Data/ImageNet/imagenet-object-localization-challenge/ILSVRC/Data/CLS-LOC"
+        ),
         batch_size=BATCH,
         num_workers=NUM_WORKERS,
         pin_memory=True,
