+++ examples/cifar_MetaFormer.py
